# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "C Unit Test Generator",
    "version": "0.1.0"
  },
  "paths": {
    "/api/project/upload": {
      "post": {
        "tags": [
          "project"
        ],
        "summary": "Upload Project",
        "operationId": "upload_project_api_project_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_project_api_project_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/{project_id}/structure": {
      "get": {
        "tags": [
          "project"
        ],
        "summary": "Get Project Structure",
        "operationId": "get_project_structure_api_project__project_id__structure_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectStructure"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/{project_id}/file": {
      "get": {
        "tags": [
          "project"
        ],
        "summary": "Get File Source",
        "operationId": "get_file_source_api_project__project_id__file_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "file_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/{project_id}/function/{function_id}": {
      "get": {
        "tags": [
          "project"
        ],
        "summary": "Get Function Detail",
        "operationId": "get_function_detail_api_project__project_id__function__function_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "function_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Function Id"
            }
          },
          {
            "name": "use_joern",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Use Joern"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_project_api_project_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_project_api_project_upload_post"
      },
...
      "FileStructure": {
        "properties": {
          "file_id": {
            "type": "string",
            "title": "File Id"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "functions": {
            "items": {
              "$ref": "#/components/schemas/FunctionInfo"
            },
            "type": "array",
            "title": "Functions"
          }
        },
        "type": "object",
        "required": [
          "file_id",
          "path",
          "functions"
        ],
        "title": "FileStructure"
      },
...
      "FunctionInfo": {
        "properties": {
          "function_id": {
            "type": "string",
            "title": "Function Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "start_line": {
            "type": "integer",
            "title": "Start Line"
          },
          "end_line": {
            "type": "integer",
            "title": "End Line"
          },
          "signature": {
            "type": "string",
            "title": "Signature"
          }
        },
        "type": "object",
        "required": [
          "function_id",
          "name",
          "start_line",
          "end_line",
          "signature"
        ],
        "title": "FunctionInfo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProjectStructure": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileStructure"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "files"
        ],
        "title": "ProjectStructure"
      },
...
      "UploadResponse": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name"
          },
          "file_count": {
            "type": "integer",
            "title": "File Count"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "project_name",
          "file_count",
          "status"
        ],
        "title": "UploadResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}